requirements.txt
faiss-cpu
sentence-transformers
langchain
langchain-ollama
ollama
numpy